Modulo5_Herramientas_para_Orquestacion/Sesion15_LangGraph_MultiAgent/Modulo6-main/requirements.txt
langchain_ollama
langchain_openai
pydantic
typeddict
langgraph
ipython
python-dotenv